refactor: split MediaInput into focused submodules
